chore(ui-react-all): remove index files from jest coverage, only allow import/export from index files
