+++ packages/configs/eslint/configs/index.js
@@ -1,6 +1,6 @@
 module.exports = {
   env: { node: true },
-  plugins: ['@typescript-eslint'],
+  plugins: ['@typescript-eslint', 'index'],
   extends: [
     'eslint:recommended',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
@@ -37,6 +37,7 @@ module.exports = {
     'generator-star-spacing': 'off',
     'global-require': 'off',
     'implicit-arrow-linebreak': 'off',
+    'index/only-import-export': 'error',
     'max-params': 2,
     'no-alert': 'error',
     'no-console': 'error',
