+++ packages/configs/eslint/package.json
@@ -19,6 +19,7 @@
     "@typescript-eslint/parser": "^5.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-index": "^1.1.5",
     "eslint-plugin-jest": "^26.1.4",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react-hooks": "^4.4.0",
