+++ packages/react-core/jest.config.js
@@ -7,9 +7,8 @@ module.exports = {
     '!<rootDir>/src/**/*(c|C)onstants.ts',
     // - __mock__ directories
     '!<rootDir>/src/**/__mock__/*',
-    // - exports files
-    '!<rootDir>/src/index.ts', // primary
-    '!<rootDir>/src/Authenticator/index.ts', // Authenticator
+    // do not collect from export files
+    '!<rootDir>/**/index.(ts|tsx)',
   ],
   coverageThreshold: {
     global: {
