+++ packages/react-core/src/types/index.ts
@@ -1,1 +1,1 @@
-export type MergeProps<C, P> = C & Omit<P, keyof C>;
+export { MergeProps } from './types';
