+++ packages/react-core/src/types/types.ts
@@ -1,0 +1,1 @@
+export type MergeProps<C, P> = C & Omit<P, keyof C>;
