+++ packages/react-geo/jest.config.js
@@ -1,14 +1,18 @@
 module.exports = {
   collectCoverage: true,
-  collectCoverageFrom: ['<rootDir>/src/**/*.(ts|tsx)'],
-  // ignore coverage for top level "export" and style files
-  coveragePathIgnorePatterns: ['<rootDir>/src/(index|styles).ts'],
+  collectCoverageFrom: [
+    '<rootDir>/src/**/*.(ts|tsx)',
+    // do not collect from export files
+    '!<rootDir>/**/index.(ts|tsx)',
+    // do not collect from top level style file
+    '!<rootDir>/src/styles.ts',
+  ],
   coverageThreshold: {
     global: {
       branches: 61,
       functions: 44,
-      lines: 70,
-      statements: 74,
+      lines: 68,
+      statements: 70,
     },
   },
   globals: { 'ts-jest': { tsconfig: 'tsconfig.json' } },
