+++ packages/react-liveness/jest.config.js
@@ -1,12 +1,16 @@
 module.exports = {
   collectCoverage: true,
-  collectCoverageFrom: ['<rootDir>/src/**/*.(ts|tsx)'],
-  // ignore coverage for top level "export", version and style files
-  coveragePathIgnorePatterns: ['<rootDir>/src/(index|styles|version).(ts|tsx)'],
+  collectCoverageFrom: [
+    '<rootDir>/src/**/*.(ts|tsx)',
+    // do not collect from export files
+    '!<rootDir>/**/index.(ts|tsx)',
+    // do not collect from top level version and styles files
+    '!<rootDir>/src/(styles|version).(ts|tsx)',
+  ],
   coverageThreshold: {
     global: {
       branches: 80,
-      functions: 85,
+      functions: 83,
       lines: 89,
       statements: 89,
     },
