+++ packages/react/src/components/ThemeProvider/index.ts
@@ -1,0 +1,6 @@
+export {
+  ColorMode,
+  Direction,
+  AmplifyProvider,
+  ThemeProvider,
+} from './ThemeProvider';
